Move and split decoration & container
* import to old modules for backward compatibility
* split constants from widget module and extend WHSettings
  (CLIP = WrapMode.CLIP, FLOW = Sizing.FLOW)
* proper normalize_*, simplify_*
  (check for None, partial handle `RELATIVE`)
