+++ urwid/__init__.py
@@ -49,33 +49,7 @@ from urwid.command_map import (
     CommandMap,
     command_map,
 )
-from urwid.container import (
-    Columns,
-    ColumnsError,
-    Frame,
-    FrameError,
-    GridFlow,
-    GridFlowError,
-    Overlay,
-    OverlayError,
-    Pile,
-    PileError,
-    WidgetContainerMixin,
-)
-from urwid.decoration import (
-    AttrMap,
-    AttrMapError,
-    AttrWrap,
-    BoxAdapter,
-    BoxAdapterError,
-    Filler,
-    FillerError,
-    Padding,
-    PaddingError,
-    WidgetDecoration,
-    WidgetDisable,
-    WidgetPlaceholder,
-)
+
 from urwid.font import (
     Font,
     FontRegistry,
@@ -139,6 +113,29 @@ from urwid.widget import (
     WidgetWrapError,
     delegate_to_widget_mixin,
     fixed_size,
+    WidgetPlaceholder,
+    AttrMap,
+    AttrMapError,
+    AttrWrap,
+    BoxAdapter,
+    BoxAdapterError,
+    WidgetDisable,
+    Filler,
+    FillerError,
+    Padding,
+    PaddingError,
+    WidgetDecoration,
+    GridFlow,
+    GridFlowError,
+    Frame,
+    FrameError,
+    Overlay,
+    OverlayError,
+    Pile,
+    PileError,
+    Columns,
+    ColumnsError,
+    WidgetContainerMixin,
 )
 from urwid.wimp import Button, CheckBox, CheckBoxError, PopUpLauncher, PopUpTarget, RadioButton, SelectableIcon
