+++ urwid/tests/test_doctests.py
@@ -9,9 +9,25 @@ import urwid.numedit
 
 def load_tests(loader, tests, ignore):
     module_doctests = [
-        urwid.widget,
         urwid.wimp,
-        urwid.decoration,
+        urwid.widget.attr_map,
+        urwid.widget.attr_wrap,
+        urwid.widget.box_adapter,
+        urwid.widget.columns,
+        urwid.widget.constants,
+        urwid.widget.container,
+        urwid.widget.divider,
+        urwid.widget.edit,
+        urwid.widget.filler,
+        urwid.widget.frame,
+        urwid.widget.grid_flow,
+        urwid.widget.overlay,
+        urwid.widget.padding,
+        urwid.widget.pile,
+        urwid.widget.solid_fill,
+        urwid.widget.text,
+        urwid.widget.widget,
+        urwid.widget.widget_decoration,
         urwid.display_common,
         urwid.event_loop.main_loop,
         urwid.numedit,
