+++ urwid/widget/__init__.py
@@ -1,31 +1,62 @@
 from __future__ import annotations
 
-from .divider import Divider
-from .edit import Edit, EditError, IntEdit
-from .solid_fill import SolidFill
-from .text import Text, TextError
-from .widget import (
+from .attr_map import AttrMap, AttrMapError
+from .attr_wrap import AttrWrap
+from .box_adapter import BoxAdapter, BoxAdapterError
+from .columns import Columns, ColumnsError
+from .constants import (
+    RELATIVE_100,
     Align,
-    BoxWidget,
-    FixedWidget,
-    FlowWidget,
     Sizing,
     VAlign,
     WHSettings,
+    WrapMode,
+    normalize_align,
+    normalize_height,
+    normalize_valign,
+    normalize_width,
+    simplify_align,
+    simplify_height,
+    simplify_valign,
+    simplify_width,
+)
+from .container import WidgetContainerMixin
+from .divider import Divider
+from .edit import Edit, EditError, IntEdit
+from .filler import Filler, FillerError, calculate_top_bottom_filler
+from .frame import Frame, FrameError
+from .grid_flow import GridFlow, GridFlowError
+from .overlay import Overlay, OverlayError
+from .padding import Padding, PaddingError, calculate_left_right_padding
+from .pile import Pile, PileError
+from .solid_fill import SolidFill
+from .text import Text, TextError
+from .widget import (
+    BoxWidget,
+    FixedWidget,
+    FlowWidget,
     Widget,
     WidgetError,
     WidgetMeta,
     WidgetWrap,
     WidgetWrapError,
-    WrapMode,
     delegate_to_widget_mixin,
     fixed_size,
     nocache_widget_render,
     nocache_widget_render_instance,
 )
+from .widget_decoration import WidgetDecoration, WidgetDisable, WidgetPlaceholder
 
 __all__ = (
     "Align",
+    "normalize_align",
+    "simplify_align",
+    "normalize_valign",
+    "simplify_valign",
+    "normalize_width",
+    "simplify_width",
+    "normalize_height",
+    "simplify_height",
     "BoxWidget",
     "Divider",
     "Edit",
@@ -67,6 +98,31 @@ __all__ = (
     "RELATIVE",
     "RELATIVE_100",
     "WEIGHT",
+    "WidgetDecoration",
+    "WidgetPlaceholder",
+    "AttrMap",
+    "AttrMapError",
+    "AttrWrap",
+    "BoxAdapter",
+    "BoxAdapterError",
+    "WidgetDisable",
+    "Padding",
+    "PaddingError",
+    "calculate_left_right_padding",
+    "Filler",
+    "FillerError",
+    "calculate_top_bottom_filler",
+    "GridFlow",
+    "GridFlowError",
+    "Overlay",
+    "OverlayError",
+    "Frame",
+    "FrameError",
+    "Pile",
+    "PileError",
+    "Columns",
+    "ColumnsError",
+    "WidgetContainerMixin",
 )
 
 # Backward compatibility
@@ -90,5 +146,4 @@ ELLIPSIS = WrapMode.ELLIPSIS
 PACK = WHSettings.PACK
 GIVEN = WHSettings.GIVEN
 RELATIVE = WHSettings.RELATIVE
-RELATIVE_100 = (WHSettings.RELATIVE, 100)
 WEIGHT = WHSettings.WEIGHT
